Update to match 0.9.0.1 semantics
As of https://github.com/meteor/meteor/commit/43e01c09eb dependencies in
the top-level .meteor/packages without an @ part have a type:
"any-reasonable" dependency which basically says "only give me a dashed
version if it was asked for explicitly in .meteor/packages or the
release or something".  (This may get more subtle in future releases.)

eg,
http://version-playground.meteor.com/#%7B%22packages%22%3A%22a%401.0.0%5Cna%401.0.1-fix%22%2C%22deps%22%3A%22a%22%7D
should not choose the dashed version, but it does without this change.